bcscratch/pyproject.toml
[tool.poetry]
name = "bcscratch"
version = "0.1.0"
description = ""
authors = ["Wei Ning (DA) <wei.ning@digitalasset.com>"]

[tool.poetry.dependencies]
python = "^3.8"
# general house-keeping
pylint = "^2.9.3"
mypy = "^0.910"
yapf = "^0.31.0"
# for note-taking
jupyter = "^1.0.0"
# needed by the project
requests = "^2.26.0"
matplotlib = "^3.4.2"
pynacl = "^1.4.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
